refactoring game.js
- game.js를 refactoring
